requirements.txt
confluent_kafka
python-dotenv
findspark
psycopg2-binary
yfinance